requirements.txt
Flask==1.1.2
pytest==6.2.2
azure-cosmosdb-table=1.0.6
